refactor(semantic): extract semantic engine into a standalone module
Move pkg/semantic and its JSON schema into a nested Go module,
github.com/bruin-data/bruin/semantic-engine, so the engine can be
consumed by lightweight tools (e.g. dac) without pulling in the full
bruin CLI dependency tree. The nested module only needs afero, yaml,
and jsonschema.

- pkg/semantic            -> semantic-engine            (package semantic, unchanged)
- schemas                 -> semantic-engine/schemas
- cmd/fetch.go now imports the nested module; the main module wires it
  via require + replace (same multi-module pattern already used by
  integration-tests/cloud-integration-tests/sail)
- make test and the CI lint job now also cover the nested module

No behavior change: the engine code and schema are byte-identical to
their previous location (moved with git rename).
