+++ .github/workflows/build-test.yml
@@ -172,3 +172,5 @@ jobs:
           curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin ${VERSION}
       - name: Run golangci-lint
         run: golangci-lint run --timeout 10m60s --build-tags="no_duckdb_arrow" ./...
+      - name: Run golangci-lint (semantic-engine module)
+        run: cd semantic-engine && golangci-lint run --timeout 10m60s ./...
