+++ go.mod
@@ -60,7 +60,7 @@ require (
 	github.com/rs/zerolog v1.28.0
 	github.com/rudderlabs/analytics-go/v4 v4.2.1
 	github.com/samber/lo v1.51.0
-	github.com/santhosh-tekuri/jsonschema/v6 v6.0.2
+	github.com/santhosh-tekuri/jsonschema/v6 v6.0.2 // indirect
 	github.com/sergi/go-diff v1.4.0
 	github.com/sijms/go-ora/v2 v2.9.0
 	github.com/snowflakedb/gosnowflake v1.17.1
@@ -291,3 +291,7 @@ require (
 	howett.net/plist v0.0.0-20181124034731-591f970eefbb // indirect
 	mvdan.cc/gofumpt v0.10.0 // indirect
 )
+
+require github.com/bruin-data/bruin/semantic-engine v0.0.0
+
+replace github.com/bruin-data/bruin/semantic-engine => ./semantic-engine
