+++ semantic-engine/schemas/embed.go
